added hash functions to Utils
